refactor: main page
- filteredIgnoreIds 상위 컴포넌트로 빼고 props로 전달
- (하위 drawer / dialog 공통으로 사용하기 위해)
- MakeQuizDrawer 컴포넌트 내부 컴포넌트 분리 (FolderDrawer / DocumentDrawer)
- SelectDrawerContent의 경우 별도 컴포넌트로 분리
- (FolderDrawer과 DocumentDrawer에서 공통으로 사용하기 위해)
- 기존 drawer-dialog로 표현하던 상위 컴포넌트 이름을 modal로 변경
